Remove driver runner without timeout
