+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/operator/DriverRunner.java
@@ -9,14 +9,11 @@ package org.elasticsearch.compute.operator;
 
 import org.elasticsearch.ExceptionsHelper;
 import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.action.support.PlainActionFuture;
 import org.elasticsearch.common.util.concurrent.CountDown;
 import org.elasticsearch.core.Releasables;
 import org.elasticsearch.tasks.TaskCancelledException;
-import org.elasticsearch.threadpool.ThreadPool;
 
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
 /**
@@ -89,34 +86,5 @@ public abstract class DriverRunner {
 
             start(driver, driverListener);
         }
-    }
-
-    /**
-     * Run all the of the listed drivers in the supplier {@linkplain ThreadPool}.
-     * @return the headers added to the context while running the drivers
-     */
-    public static Map<String, List<String>> runToCompletion(ThreadPool threadPool, int maxIterations, List<Driver> drivers) {
-        DriverRunner runner = new DriverRunner() {
-            @Override
-            protected void start(Driver driver, ActionListener<Void> driverListener) {
-                Driver.start(threadPool.executor("esql"), driver, maxIterations, driverListener);
-            }
-        };
-        AtomicReference<Map<String, List<String>>> responseHeaders = new AtomicReference<>();
-        PlainActionFuture<Void> future = new PlainActionFuture<>();
-        runner.runToCompletion(drivers, new ActionListener<>() {
-            @Override
-            public void onResponse(Void unused) {
-                responseHeaders.set(threadPool.getThreadContext().getResponseHeaders());
-                future.onResponse(null);
-            }
-
-            @Override
-            public void onFailure(Exception e) {
-                future.onFailure(e);
-            }
-        });
-        future.actionGet();
-        return responseHeaders.get();
     }
 }
